+++ test/resources/support/test_countable.rb
@@ -5,7 +5,7 @@ require 'test_helper'
 describe KlaviyoAPI::Support::Countable do
   before do
     stub_request(:get, "https://a.klaviyo.com/api/v1/people?api_key=&count=1")
-      .to_return body: load_fixture(:count)
+      .to_return body: load_fixture(:paginated_collection)
   end
 
   it 'successfully gets a count' do